# api/openapi.yaml
openapi: 3.1.0
info:
  title: BTG LMS application API
  version: 0.1.0
  description: Same-origin session transport API, public immutable published-course read API, and authenticated private Authoring reads and mutations. Authoring responses are no-store. Every Authoring mutation requires a trusted Origin and the authenticated session-bound CSRF token; unauthorized resources are hidden as 404.
paths:
  /api/authoring/drafts:
    get:
      operationId: listAuthoringDrafts
      description: Returns minimal mutable Draft summaries for which the authenticated actor currently has an active Authoring membership. Revoked memberships are excluded. Responses are private and not cacheable.
      security: [{ SessionCookie: [] }]
      responses:
        '200': { description: Accessible Draft summaries ordered by most recently updated, then ID, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringDraftList' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
    post:
      operationId: createAuthoringDraft
      description: Creates a new editable Draft for the authenticated actor. The creator receives the initial active MAINTAINER membership atomically. The server creates the opaque Course identity; no Review or CourseVersion is created.
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringOrigin' }, { $ref: '#/components/parameters/AuthoringCSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthoringDraftCreateRequest' }
      responses:
        '201': { description: Created editable Draft, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringDraft' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /api/authoring/drafts/{draftId}:
    get:
      operationId: getAuthoringDraft
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }]
      responses:
        '200': { description: Draft metadata and current revision, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringDraft' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
    patch:
      operationId: updateAuthoringDraftMetadata
      description: Updates only mutable draft metadata using an explicit optimistic-concurrency revision. The request requires the authenticated session's CSRF token and a trusted same-origin Origin. Responses are private and not cacheable.
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }, { $ref: '#/components/parameters/AuthoringOrigin' }, { $ref: '#/components/parameters/AuthoringCSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthoringDraftUpdateRequest' }
      responses:
        '200': { description: Committed updated draft metadata with incremented revision, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringDraft' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /api/authoring/drafts/{draftId}/workspace:
    get:
      operationId: getAuthoringWorkspace
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }]
      responses:
        '200': { description: Private workspace metadata, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringWorkspace' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /api/authoring/drafts/{draftId}/assets:
    get:
      operationId: listAuthoringDraftAssets
      description: Returns a bounded, newest-first page of AVAILABLE Assets owned by the exact authorized Draft. It returns no binary URL or private provenance.
      security: [{ SessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/AuthoringDraftID' }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: offset, in: query, required: false, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        '200': { description: Page of safe Draft Asset summaries, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringAssetList' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
    post:
      operationId: uploadAuthoringAsset
      description: Streams exactly one file into the authorized Draft. Active AUTHOR and MAINTAINER members have authoring.asset.upload. The server derives Draft ownership and creator identity, detects media type from bytes, measures size and SHA-256, and returns only safe AVAILABLE Asset metadata.
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }, { $ref: '#/components/parameters/AuthoringOrigin' }, { $ref: '#/components/parameters/AuthoringCSRFToken' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201': { description: Authoritative AVAILABLE Asset metadata suitable for a canonical assetKey reference, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringAsset' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '413': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /api/authoring/drafts/{draftId}/modules:
    post:
      operationId: createAuthoringModule
      description: Creates one draft Module at a zero-based position using the current draft revision. The authenticated session must provide a valid CSRF token and trusted same-origin Origin.
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }, { $ref: '#/components/parameters/AuthoringOrigin' }, { $ref: '#/components/parameters/AuthoringCSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthoringModuleCreateRequest' }
      responses:
        '200': { description: Committed Module and current Draft revision, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringModuleMutationResponse' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /api/authoring/drafts/{draftId}/members:
    get:
      operationId: listAuthoringActiveMembers
      description: Returns current active AUTHOR and MAINTAINER memberships for one Draft. The authenticated actor must have authoring.read. Revoked membership history is not exposed.
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }]
      responses:
        '200': { description: Active opaque Draft memberships ordered by user ID, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringActiveMemberList' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
    post:
      operationId: addAuthoringMember
      description: Adds an active AUTHOR or MAINTAINER membership using the Draft revision. The authenticated actor must have authoring.members.manage.
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }, { $ref: '#/components/parameters/AuthoringOrigin' }, { $ref: '#/components/parameters/AuthoringCSRFToken' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringMemberAddRequest' } } }
      responses:
        '200': { description: Committed membership and current Draft revision, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringMemberMutationResponse' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /api/authoring/drafts/{draftId}/members/{userId}:
    patch:
      operationId: changeAuthoringMemberRole
      description: Replaces one active member role. Demoting the final active MAINTAINER is rejected.
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }, { $ref: '#/components/parameters/AuthoringUserID' }, { $ref: '#/components/parameters/AuthoringOrigin' }, { $ref: '#/components/parameters/AuthoringCSRFToken' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringMemberRoleRequest' } } }
      responses:
        '200': { description: Committed active membership and current Draft revision, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringMemberMutationResponse' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: revokeAuthoringMember
      description: Revokes one active membership while preserving history. Revoking the final active MAINTAINER is rejected.
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }, { $ref: '#/components/parameters/AuthoringUserID' }, { $ref: '#/components/parameters/AuthoringOrigin' }, { $ref: '#/components/parameters/AuthoringCSRFToken' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringMemberRevokeRequest' } } }
      responses:
        '200': { description: Revoked membership and current Draft revision, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringMemberMutationResponse' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /api/authoring/drafts/{draftId}/modules/order:
    put:
      operationId: reorderAuthoringModules
      description: Atomically replaces the complete ordered Module ID list for one Draft. The list must contain every current Module exactly once.
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }, { $ref: '#/components/parameters/AuthoringOrigin' }, { $ref: '#/components/parameters/AuthoringCSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthoringModuleReorderRequest' }
      responses:
        '200': { description: Committed Draft revision, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringModuleOrderResponse' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /api/authoring/drafts/{draftId}/modules/{moduleId}:
    patch:
      operationId: updateAuthoringModule
      description: Updates title and description only. Stable key and position are immutable through PATCH.
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }, { $ref: '#/components/parameters/AuthoringModuleID' }, { $ref: '#/components/parameters/AuthoringOrigin' }, { $ref: '#/components/parameters/AuthoringCSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthoringModuleUpdateRequest' }
      responses:
        '200': { description: Committed Module and current Draft revision, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringModuleMutationResponse' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: deleteAuthoringModule
      description: Deletes an empty Module only. Modules containing Lessons are rejected so the lesson cascade remains unavailable through this API.
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }, { $ref: '#/components/parameters/AuthoringModuleID' }, { $ref: '#/components/parameters/AuthoringOrigin' }, { $ref: '#/components/parameters/AuthoringCSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthoringModuleDeleteRequest' }
      responses:
        '200': { description: Committed Draft revision, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringModuleOrderResponse' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /api/authoring/drafts/{draftId}/modules/{moduleId}/lessons:
    post:
      operationId: createAuthoringLesson
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }, { $ref: '#/components/parameters/AuthoringModuleID' }, { $ref: '#/components/parameters/AuthoringOrigin' }, { $ref: '#/components/parameters/AuthoringCSRFToken' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringLessonCreateRequest' } } }
      responses:
        '200': { description: Committed Lesson metadata and current Draft revision, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringLessonMutationResponse' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
...
  /api/authoring/drafts/{draftId}/structure:
    get:
      operationId: getAuthoringDraftStructure
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AuthoringDraftID' }]
      responses:
        '200': { description: Ordered modules and lesson summaries without full lesson content, content: { application/json: { schema: { $ref: '#/components/schemas/AuthoringStructure' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
    AuthoringOrigin: { name: Origin, in: header, required: true, schema: { type: string }, description: Trusted application Origin required for browser mutations. }
    AuthoringCSRFToken: { name: X-CSRF-Token, in: header, required: true, schema: { type: string }, description: CSRF token bound to the authenticated session. }
    AuthoringDraftID: { name: draftId, in: path, required: true, schema: { type: string, format: uuid } }
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: btg_session
      description: Opaque server-side session token; HttpOnly, SameSite=Lax, Secure in production.
  responses:
    Problem:
      description: RFC 9457 Problem Details with request correlation ID
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    AuthoringDraftList:
      type: object
      additionalProperties: false
      required: [drafts]
      properties:
        drafts: { type: array, items: { $ref: '#/components/schemas/AuthoringDraftSummary' } }
    AuthoringWorkspace:
      type: object
      required: [id, draft_id, created_at, last_activity_at]
      properties:
        id: { type: string, format: uuid }
        draft_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        last_activity_at: { type: string, format: date-time }
...